feat:finish ac4

+++ src/main/java/org/example/todolistbackend/service/TodoService.java
@@ -35,7 +35,11 @@ public class TodoService {
                 .orElse(null);
     }
     
-    public void deleteTodo(Long id) {
-        todoRepository.deleteById(id);
+    public boolean deleteTodo(Long id) {
+        if (todoRepository.existsById(id)) {
+            todoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
